requirements.txt
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.1.0
six==1.17.0
tqdm==4.67.1
tzdata==2024.2
wheel==0.44.0
